\name{sentiment_afinn}
\docType{data}
\alias{sentiment_afinn}
\title{AFINN Sentiment Lexicon}
\description{
The AFINN lexicon is a list of English terms manually rated
for valence with an integer between -5 (negative) and +5
(positive) by Finn \enc{Årup}{Aarup} Nielsen between
2009 and 2011.

The original lexicon contains some multi-word phrases, but they
are excluded here.

The original lexicon is distributed under the
\href{https://opendatacommons.org/licenses/odbl/1-0/}{Open Database License (ODbL) v1.0}.
You are free to share, create works from, and adapt the lexicon, as long as you
attribute the original lexicon in your work. If you adapt the lexicon, you
must keep the adapted lexicon open and apply a similar license.
}
\usage{sentiment_afinn}
\format{A data frame with one row for each term}
\source{\url{http://www2.imm.dtu.dk/pubdb/views/publication_details.php?id=6010}}
\references{
Finn \enc{Årup}{Aarup} Nielsen
A new ANEW: Evaluation of a word list for sentiment analysis in microblogs.
\emph{Proceedings of the ESWC2011 Workshop on 'Making Sense of Microposts': Big things come in small packages 718 in CEUR Workshop Proceedings}
93-98. 2011 May.
\url{http://arxiv.org/abs/1103.2903}.
}
\keyword{datasets}
