\name{corpus-deprecated}
\alias{corpus-deprecated}
\title{Deprecated Functions in Package \pkg{corpus}}
\description{
These functions are provided for compatibility with older versions of
\pkg{corpus} only, and may be defunct as soon as the next release.
}
%\usage{
%}
%\arguments{
%}
%\details{
%}
\seealso{
\code{\link{Deprecated}}
}
\keyword{internal}
\keyword{misc}
